+++ src/main/frontend/pipeline-console-view/pipeline-console/main/ConsoleLogCard.tsx
@@ -18,6 +18,7 @@ import {
   StepInfo,
   StepLogBufferInfo,
 } from "./PipelineConsoleModel.tsx";
+import InputStep from "./steps/InputStep.tsx";
 
 const ConsoleLogStream = lazy(() => import("./ConsoleLogStream.tsx"));
 
@@ -45,56 +46,7 @@ export default function ConsoleLogCard(props: ConsoleLogCardProps) {
 
   const inputStep = props.step.inputStep;
   if (inputStep && !inputStep.parameters) {
-    function handler(id: string, action: string) {
-      fetch(`../input/${id}/${action}`, {
-        method: "POST",
-        headers: window.crumb.wrap({}),
-      })
-        .then((res) => {
-          if (!res.ok) {
-            console.error(res);
-          }
-          return true;
-        })
-        .catch((err) => {
-          console.error(err);
-        });
-    }
-
-    const ok = () => {
-      return handler(inputStep.id, "proceedEmpty");
-    };
-
-    const abort = () => {
-      return handler(inputStep.id, "abort");
-    };
-
-    return (
-      <div className="pgv-input-step">
-        <div className="pgv-step-detail-header__content">
-          <StatusIcon
-            status={props.step.state}
-            percentage={props.step.completePercent}
-          />
-          <span>{inputStep.message}</span>
-        </div>
-        <div
-          className={
-            "jenkins-buttons-row jenkins-buttons-row--equal-width pgv-input-step__controls"
-          }
-        >
-          <button
-            onClick={ok}
-            className={"jenkins-button jenkins-button--primary"}
-          >
-            {inputStep.ok}
-          </button>
-          <button onClick={abort} className={"jenkins-button"}>
-            {inputStep.cancel}
-          </button>
-        </div>
-      </div>
-    );
+    return <InputStep {...props} />;
   }
 
   return (
@@ -193,16 +145,6 @@ export default function ConsoleLogCard(props: ConsoleLogCardProps) {
       )}
     </div>
   );
-}
-
-declare global {
-  interface Window {
-    crumb: Crumb;
-  }
-
-  interface Crumb {
-    wrap: (headers: Record<string, string>) => Record<string, string>;
-  }
 }
 
 function ConsoleLogBody(props: ConsoleLogCardProps) {
